requirements.txt
pandas
numpy
scikit-learn
joblib
nltk
spacy
pdfminer.six
python-docx
tqdm
streamlit
pyyaml